worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.6.3",
    "firecrawl-py>=4.22.0",
    "langchain>=1.2.15",
    "langchain-groq>=1.1.2",
    "pydantic-settings>=2.13.1",
    "python-socketio>=5.16.1",
    "sqlalchemy>=2.0.44",
    "psycopg[binary]>=3.2.11",
]
